--- src/modules/fancyzones/FancyZonesLib/ZoneColors.h
@@ -1,11 +1,0 @@
-#pragma once
-#include <windef.h>
-
-struct ZoneColors
-{
-    COLORREF primaryColor;
-    COLORREF borderColor;
-    COLORREF highlightColor;
-    COLORREF numberColor;
-    int highlightOpacity;
-};
